fix initialization order, remove circular dependency
